update - update migrations files

+++ database/migrations/2019_09_08_222307_create_mbs_messages_files_table.php
@@ -22,7 +22,7 @@ class CreateMbsMessagesFilesTable extends Migration
             $table->string('file_name_old');
             $table->string('file_path', 100);
             $table->string('video_thumbnail', 100);
-            $table->integer('user_id');
+            $table->bigInteger('user_id');
             $table->timestamps();
             $table->softDeletes();
         });
